Handle Monzo split transactions (#36)
When you split transactions in the Monzo App, the transaction from
their API contains a CounterParty AND a Merchant. In the current
logic, the Merchant name is used which leads to confusion because
in the App you see that the transaction is to the CounterParty.

Hence, we should use the CounterParty name, and if not present,
override the domain.Transaction notes with the Merchant name.